Add PC port infrastructure using PsyCross (SDL2 + OpenGL)
Port the Silent Hill PS1 decompilation to PC using PsyCross as a PSX
hardware abstraction layer. The game boots through all startup screens
(Boot -> Konami -> KCET -> MovieIntro -> MainMenu) with visual output
rendering to an SDL window.

Key components:
- CMake build system targeting MSYS2/MinGW64
- PSY-Q header shims bridging to PsyCross equivalents
- PSX memory emulation (2MB RAM, scratchpad, address translation)
- GS library stubs implementing GsInitGraph, GsDrawOt, etc.
- Sound/pad/API stubs for unimplemented subsystems
- Map overlay stubs (statically linked instead of dynamic loading)

Rendering fixes:
- Disable PGXP (VERTTYPE half/short mismatch causes NaN vertices)
- Force dfe=1 in PsyCross (PSX dfe flag misused as render target selector)
- Eliminate PSX VRAM double-buffer offsets (draw env offset pushed content off-screen)
- Center coordinate system (set draw offset to screen_w/2, screen_h/2)
- Fix P_TAG/setCodeWord for 64-bit (12-byte P_TAG vs PSX 4-byte)

Runtime fixes for 64-bit compatibility:
- PSX_ADDR macro for translating PSX addresses to emulated RAM
- uintptr_t casts replacing s32 pointer truncation
- Screen fade acceleration (4x speed on PC)
- Skip hardware-dependent init (vibration, memory cards)
- Delta time injection in internal loops

+++ src/bodyprog/dms.c
@@ -27,8 +27,8 @@ void DmsHeader_FixOffsets(s_DmsHeader* dmsHdr) // 0x8008C9A0
     dmsHdr->isLoaded_0 = true;
 
     // Add memory address of DMS header to offsets in `dmsHdr`.
-    dmsHdr->intervalPtr_8 = (u8*)dmsHdr->intervalPtr_8 + (u32)dmsHdr;
-    dmsHdr->characters_18 = (u8*)dmsHdr->characters_18 + (u32)dmsHdr;
+    dmsHdr->intervalPtr_8 = (u8*)dmsHdr->intervalPtr_8 + (uintptr_t)dmsHdr;
+    dmsHdr->characters_18 = (u8*)dmsHdr->characters_18 + (uintptr_t)dmsHdr;
 
     DmsEntry_FixOffsets(&dmsHdr->camera_1C, dmsHdr);
 
@@ -42,8 +42,8 @@ void DmsHeader_FixOffsets(s_DmsHeader* dmsHdr) // 0x8008C9A0
 
 void DmsEntry_FixOffsets(s_DmsEntry* entry, s_DmsHeader* dmsHdr) // 0x8008CA44
 {
-    entry->keyframes_C.character = (u32)entry->keyframes_C.character + (u32)dmsHdr;
-    entry->svectorPtr_8          = (u32)entry->svectorPtr_8 + (u32)dmsHdr;
+    entry->keyframes_C.character = (u32)entry->keyframes_C.character + (uintptr_t)dmsHdr;
+    entry->svectorPtr_8          = (u32)entry->svectorPtr_8 + (uintptr_t)dmsHdr;
 }
 
 s_DmsInterval* func_8008CA60(volatile s32 unused, s32 idx, s_DmsHeader* dmsHdr) // 0x8008CA60
